Add files via upload
This repository contains a Hotel Management System implemented in Assembly Language for the EMU 8086 processor. The project demonstrates fundamental concepts of assembly programming, including data handling, control structures. The system allows users to perform essential operations such as dinner menu, breakfast menu etc.

Explore the code to understand low-level programming techniques and the architecture of the 8086 microprocessor. This project serves as a practical example for those interested in assembly language and systems programming.